Add Audit functionilaity
And cleaning code a little bit in reg_oracle_fict
